refactor: split oversized modules into focused submodules
