+++ src/subtitle_ocr/text_processing/english_words.rs
@@ -1,0 +1,516 @@
+pub(super) fn is_common_english_word(word: &str) -> bool {
+    const WORDS: [&str; 511] = [
+        "a",
+        "abandon",
+        "able",
+        "about",
+        "above",
+        "according",
+        "across",
+        "actually",
+        "add",
+        "after",
+        "again",
+        "against",
+        "ago",
+        "ahead",
+        "air",
+        "alive",
+        "all",
+        "allow",
+        "almost",
+        "along",
+        "already",
+        "also",
+        "although",
+        "always",
+        "am",
+        "among",
+        "amount",
+        "an",
+        "and",
+        "another",
+        "any",
+        "anyone",
+        "anything",
+        "apostatize",
+        "apostatized",
+        "are",
+        "around",
+        "as",
+        "ask",
+        "asked",
+        "at",
+        "away",
+        "back",
+        "bad",
+        "base",
+        "be",
+        "became",
+        "because",
+        "become",
+        "becoming",
+        "been",
+        "before",
+        "beg",
+        "behind",
+        "being",
+        "below",
+        "beside",
+        "best",
+        "better",
+        "between",
+        "black",
+        "blood",
+        "blue",
+        "body",
+        "book",
+        "born",
+        "both",
+        "break",
+        "bring",
+        "brother",
+        "brought",
+        "build",
+        "built",
+        "burn",
+        "burning",
+        "but",
+        "by",
+        "call",
+        "came",
+        "can",
+        "cannot",
+        "cause",
+        "century",
+        "children",
+        "christ",
+        "christi",
+        "christians",
+        "church",
+        "city",
+        "clear",
+        "close",
+        "coal",
+        "cold",
+        "come",
+        "constable",
+        "continue",
+        "could",
+        "country",
+        "courage",
+        "crane",
+        "cut",
+        "dandelion",
+        "dark",
+        "day",
+        "days",
+        "dead",
+        "death",
+        "deep",
+        "demonstrate",
+        "denounced",
+        "despite",
+        "did",
+        "die",
+        "different",
+        "do",
+        "does",
+        "done",
+        "door",
+        "down",
+        "drops",
+        "during",
+        "dutch",
+        "each",
+        "early",
+        "earth",
+        "easy",
+        "either",
+        "ended",
+        "enough",
+        "even",
+        "ever",
+        "every",
+        "face",
+        "fact",
+        "faith",
+        "family",
+        "far",
+        "father",
+        "fear",
+        "ferreira",
+        "few",
+        "figure",
+        "filled",
+        "finally",
+        "fire",
+        "first",
+        "five",
+        "floor",
+        "follow",
+        "food",
+        "for",
+        "force",
+        "form",
+        "four",
+        "free",
+        "friars",
+        "friend",
+        "from",
+        "front",
+        "full",
+        "game",
+        "get",
+        "girl",
+        "give",
+        "given",
+        "gives",
+        "glass",
+        "go",
+        "god",
+        "good",
+        "gospel",
+        "governor",
+        "great",
+        "green",
+        "group",
+        "grow",
+        "had",
+        "half",
+        "hand",
+        "happened",
+        "hard",
+        "has",
+        "have",
+        "he",
+        "head",
+        "hear",
+        "heart",
+        "heavy",
+        "hells",
+        "her",
+        "here",
+        "hes",
+        "hidden",
+        "him",
+        "his",
+        "history",
+        "holes",
+        "home",
+        "hope",
+        "hot",
+        "house",
+        "how",
+        "however",
+        "i",
+        "idea",
+        "if",
+        "important",
+        "in",
+        "inside",
+        "into",
+        "is",
+        "it",
+        "its",
+        "japan",
+        "japanese",
+        "jesus",
+        "job",
+        "just",
+        "keep",
+        "kind",
+        "king",
+        "knew",
+        "know",
+        "known",
+        "ladles",
+        "land",
+        "large",
+        "last",
+        "late",
+        "later",
+        "leave",
+        "left",
+        "let",
+        "letter",
+        "life",
+        "light",
+        "like",
+        "line",
+        "list",
+        "little",
+        "live",
+        "living",
+        "long",
+        "looked",
+        "looking",
+        "looks",
+        "lopped",
+        "lord",
+        "lost",
+        "love",
+        "low",
+        "made",
+        "main",
+        "make",
+        "man",
+        "many",
+        "may",
+        "me",
+        "mean",
+        "means",
+        "men",
+        "might",
+        "mind",
+        "minute",
+        "mockery",
+        "money",
+        "month",
+        "months",
+        "more",
+        "morning",
+        "most",
+        "mother",
+        "mountain",
+        "move",
+        "moved",
+        "must",
+        "my",
+        "nagasaki",
+        "name",
+        "near",
+        "need",
+        "never",
+        "new",
+        "news",
+        "next",
+        "night",
+        "no",
+        "none",
+        "nor",
+        "north",
+        "not",
+        "nothing",
+        "notice",
+        "now",
+        "number",
+        "of",
+        "off",
+        "officials",
+        "often",
+        "old",
+        "on",
+        "once",
+        "one",
+        "only",
+        "open",
+        "or",
+        "order",
+        "other",
+        "others",
+        "our",
+        "out",
+        "outside",
+        "over",
+        "own",
+        "padres",
+        "pain",
+        "paper",
+        "pardon",
+        "part",
+        "partly",
+        "pass",
+        "past",
+        "pax",
+        "pay",
+        "peace",
+        "people",
+        "persecution",
+        "person",
+        "piece",
+        "place",
+        "point",
+        "portugal",
+        "possible",
+        "power",
+        "praise",
+        "praised",
+        "presence",
+        "priests",
+        "probably",
+        "progress",
+        "prolonged",
+        "proven",
+        "public",
+        "put",
+        "question",
+        "rain",
+        "ransomed",
+        "rather",
+        "reach",
+        "read",
+        "ready",
+        "really",
+        "red",
+        "refused",
+        "remain",
+        "remained",
+        "remember",
+        "repression",
+        "right",
+        "risked",
+        "river",
+        "road",
+        "room",
+        "run",
+        "said",
+        "same",
+        "saw",
+        "say",
+        "school",
+        "second",
+        "secret",
+        "see",
+        "seem",
+        "seen",
+        "service",
+        "set",
+        "several",
+        "shall",
+        "she",
+        "show",
+        "side",
+        "since",
+        "sir",
+        "six",
+        "slowly",
+        "small",
+        "smuggled",
+        "so",
+        "society",
+        "some",
+        "something",
+        "son",
+        "soon",
+        "sound",
+        "south",
+        "splash",
+        "spread",
+        "springs",
+        "stand",
+        "state",
+        "still",
+        "stop",
+        "stopped",
+        "story",
+        "strength",
+        "stronger",
+        "strongest",
+        "such",
+        "suffering",
+        "sure",
+        "surrendered",
+        "sweeping",
+        "table",
+        "taken",
+        "taking",
+        "talk",
+        "teacher",
+        "team",
+        "tell",
+        "ten",
+        "terms",
+        "text",
+        "than",
+        "that",
+        "thats",
+        "the",
+        "their",
+        "them",
+        "then",
+        "there",
+        "these",
+        "they",
+        "thing",
+        "things",
+        "think",
+        "this",
+        "those",
+        "though",
+        "thousands",
+        "three",
+        "through",
+        "time",
+        "to",
+        "today",
+        "together",
+        "told",
+        "tonight",
+        "too",
+        "took",
+        "top",
+        "tortured",
+        "toward",
+        "trader",
+        "traveling",
+        "true",
+        "truth",
+        "turn",
+        "two",
+        "under",
+        "until",
+        "unzen",
+        "up",
+        "upon",
+        "us",
+        "use",
+        "used",
+        "using",
+        "usually",
+        "value",
+        "very",
+        "view",
+        "voice",
+        "wait",
+        "wall",
+        "want",
+        "war",
+        "was",
+        "watch",
+        "water",
+        "we",
+        "week",
+        "well",
+        "went",
+        "were",
+        "what",
+        "when",
+        "where",
+        "which",
+        "white",
+        "who",
+        "whole",
+        "whose",
+        "why",
+        "will",
+        "with",
+        "within",
+        "woman",
+        "word",
+        "words",
+        "work",
+        "world",
+        "worse",
+        "would",
+        "wrote",
+        "year",
+        "years",
+        "yes",
+        "yet",
+        "you",
+        "young",
+        "your",
+    ];
+    WORDS.binary_search(&word).is_ok()
+}
